// src/store/cms/cms.enum.ts
enum CmsActionTypeEnum {
	GET_CMS_SUCCESS = "GET_CMS_SUCCESS",
	GET_CMS_PENDING = "GET_CMS_PENDING",
	GET_CMS_FAILED = "GET_CMS_FAILED",

	GET_SINGLE_CMS_SUCCESS = "GET_SINGLECMS_SUCCESS",
	GET_SINGLE_CMS_PENDING = "GET_SINGLE_CMS_PENDING",
	GET_SINGLE_CMS_FAILED = "GET_SINGLE_CMS_FAILED",

	UPDATE_CMS_SUCCESS = "UPDATE_CMS_SUCCESS",
	UPDATE_CMS_PENDING = "UPDATE_CMS_PENDING",
	UPDATE_CMS_FAILED = "UPDATE_CMS_FAILED",
}
export default CmsActionTypeEnum;